requirements.txt
scipy==1.2.1
scikit-learn
opencv-python
h5py
matplotlib
Pillow
requests
psutil

